+++ src/main/java/damon/backend/exception/custom/AccessTokenNotFoundException.java
@@ -6,6 +6,6 @@ import damon.backend.exception.Status;
 public class AccessTokenNotFoundException extends CustomException {
 
     public AccessTokenNotFoundException() {
-        super(404, Status.NOT_FOUND, "토큰을 찾을 수 없습니다.");
+        super(404, Status.NOT_FOUND, "엑세스 토큰을 찾을 수 없습니다.");
     }
 }
