+++ src/main/java/damon/backend/exception/custom/RefreshTokenNotFoundException.java
@@ -6,6 +6,6 @@ import damon.backend.exception.Status;
 public class RefreshTokenNotFoundException extends CustomException {
 
     public RefreshTokenNotFoundException() {
-        super(404, Status.NOT_FOUND, "토큰을 찾을 수 없습니다.");
+        super(404, Status.NOT_FOUND, "리프레시 토큰을 찾을 수 없습니다.");
     }
 }
